fix(db): resolve ambiguous generate_series(smallint,smallint) in seat allocator
Every numbered (Bar) booking 500'd at allocate_seats_or_throw() because
generate_series(v_start, v_end) was called with smallint bounds; PostgreSQL
has no (smallint,smallint) overload and the implicit cast to int4/int8/numeric
is ambiguous ('function generate_series(smallint, smallint) is not unique').

Cast the bounds to ::int so it resolves to generate_series(int,int); the
result is still cast back to smallint[] (seat numbers fit in smallint, lossless).
Latent since 0011, carried through 0015/0022/0024. Adds forward migration 0025
(CREATE OR REPLACE, signature unchanged) and fixes the inline copies in the
combined SUPABASE_RUN_* paste files. Applied to prod Supabase; booking verified 201.

+++ SUPABASE_RUN_THIS.sql
@@ -732,7 +732,7 @@ begin
         if v_seat = any(v_taken) then v_ok := false; exit; end if;
       end loop;
       if v_ok then
-        v_result := array(select generate_series(v_start, v_end))::smallint[];
+        v_result := array(select generate_series(v_start::int, v_end::int))::smallint[];
         exit;
       end if;
       v_end := v_end - 1;
@@ -858,7 +858,7 @@ begin
       end if;
     end loop;
     if v_ok then
-      v_result := array(select generate_series(v_start, v_end))::smallint[];
+      v_result := array(select generate_series(v_start::int, v_end::int))::smallint[];
       return v_result;
     end if;
     v_end := v_end - 1;
@@ -1308,7 +1308,7 @@ begin
       if v_seat = any(v_taken) then v_ok := false; exit; end if;
     end loop;
     if v_ok then
-      v_result := array(select generate_series(v_start, v_end))::smallint[];
+      v_result := array(select generate_series(v_start::int, v_end::int))::smallint[];
       return v_result;
     end if;
     v_end := v_end - 1;
